Repackage client-blaze to blaze-client
