+++ blaze-client/src/main/scala/org/http4s/blaze/client/BlazeClient.scala
@@ -15,8 +15,8 @@
  */
 
 package org.http4s
-package client
 package blaze
+package client
 
 import cats.effect._
 import cats.effect.concurrent._
@@ -26,6 +26,7 @@ import java.nio.ByteBuffer
 import java.util.concurrent.TimeoutException
 import org.http4s.blaze.util.TickWheelExecutor
 import org.http4s.blazecore.{IdleTimeoutStage, ResponseHeaderTimeoutStage}
+import org.http4s.client.{Client, RequestKey}
 import org.log4s.getLogger
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
