+++ blaze-client/src/main/scala/org/http4s/blaze/client/Http1Client.scala
@@ -15,14 +15,14 @@
  */
 
 package org.http4s
-package client
 package blaze
+package client
 
 import cats.effect._
 import fs2.Stream
 import org.http4s.blaze.channel.ChannelOptions
+import org.http4s.client.{Client, ConnectionBuilder}
 import org.http4s.internal.SSLContextOption
-
 import scala.concurrent.duration.Duration
 
 /** Create a HTTP1 client which will attempt to recycle connections */
