+++ blaze-client/src/main/scala/org/http4s/blaze/client/PoolManager.scala
@@ -15,13 +15,14 @@
  */
 
 package org.http4s
-package client
 package blaze
+package client
 
 import cats.effect._
 import cats.effect.concurrent.Semaphore
 import cats.syntax.all._
 import java.time.Instant
+import org.http4s.client.{Connection, ConnectionBuilder, RequestKey}
 import org.log4s.getLogger
 import scala.collection.mutable
 import scala.concurrent.ExecutionContext
