+++ blaze-client/src/test/scala-2.13/org/http4s/client/blaze/BlazeClient213Suite.scala
@@ -22,6 +22,7 @@ import cats.effect.concurrent.Ref
 import cats.syntax.all._
 import fs2.Stream
 import org.http4s._
+import org.http4s.blaze.client.BlazeClientBase
 import scala.concurrent.duration._
 import scala.util.Random
