+++ blaze-client/src/test/scala/org/http4s/blaze/client/BlazeHttp1ClientSuite.scala
@@ -15,26 +15,15 @@
  */
 
 package org.http4s
-package client
 package blaze
+package client
 
 import cats.effect.IO
-import java.nio.ByteBuffer
-import org.http4s.blaze.pipeline.{HeadStage, LeafBuilder}
+import org.http4s.client.ClientRouteTestBattery
+import org.http4s.internal.threads.newDaemonPoolExecutionContext
 
-private[blaze] object MockClientBuilder {
-  def builder(
-      head: => HeadStage[ByteBuffer],
-      tail: => BlazeConnection[IO]): ConnectionBuilder[IO, BlazeConnection[IO]] = { _ =>
-    IO {
-      val t = tail
-      LeafBuilder(t).base(head)
-      t
-    }
-  }
-
-  def manager(
-      head: => HeadStage[ByteBuffer],
-      tail: => BlazeConnection[IO]): ConnectionManager[IO, BlazeConnection[IO]] =
-    ConnectionManager.basic(builder(head, tail))
+class BlazeHttp1ClientSuite extends ClientRouteTestBattery("BlazeClient") {
+  def clientResource =
+    BlazeClientBuilder[IO](
+      newDaemonPoolExecutionContext("blaze-pooled-http1-client-spec", timeout = true)).resource
 }
