+++ examples/blaze/src/main/scala/com/example/http4s/blaze/ClientExample.scala
@@ -22,7 +22,7 @@ import org.http4s.Status.{NotFound, Successful}
 import org.http4s.circe._
 import org.http4s.syntax.all._
 import org.http4s.client.Client
-import org.http4s.client.blaze.BlazeClientBuilder
+import org.http4s.blaze.client.BlazeClientBuilder
 import scala.concurrent.ExecutionContext.global
 
 object ClientExample extends IOApp {
