+++ examples/blaze/src/main/scala/com/example/http4s/blaze/demo/client/MultipartClient.scala
@@ -22,7 +22,7 @@ import fs2.Stream
 import java.net.URL
 import org.http4s._
 import org.http4s.Method._
-import org.http4s.client.blaze.BlazeClientBuilder
+import org.http4s.blaze.client.BlazeClientBuilder
 import org.http4s.client.dsl.Http4sClientDsl
 import org.http4s.headers.`Content-Type`
 import org.http4s.implicits._
