+++ scalafix/build.sbt
@@ -59,7 +59,8 @@ lazy val input = project.settings(
       "http4s-jetty",
       "http4s-jetty-client",
       "http4s-okhttp-client",
-      "http4s-async-http-client"
+      "http4s-async-http-client",
+      "http4s-blaze-client"
     ).map("org.http4s" %% _ % "0.21.18"),
   skip in publish := true
 )
