+++ scalafix/rules/src/main/scala/fix/RewritePackages.scala
@@ -26,6 +26,7 @@ class RewritePackages extends SemanticRule("RewritePackages") {
       "org.http4s.server.jetty" -> "org.http4s.jetty.server",
       "org.http4s.client.jetty" -> "org.http4s.jetty.client",
       "org.http4s.client.okhttp" -> "org.http4s.okhttp.client",
-      "org.http4s.client.asynchttpclient" -> "org.http4s.asynchttpclient.client"
+      "org.http4s.client.asynchttpclient" -> "org.http4s.asynchttpclient.client",
+      "org.http4s.client.blaze" -> "org.http4s.blaze.client"
     )
 }
